+++ src/compiler/binary.ts
@@ -23,6 +23,8 @@ import {
 	shouldPushToPrecedingStatement,
 } from "../utility/type";
 import { compileBindingLiteral, getSubTypeOrThrow } from "./binding";
+import { isDefinedAsMethod } from "./call";
+import { isMethodDeclaration } from "./function";
 import { compileLogicalBinary } from "./truthiness";
 
 function getLuaAddExpression(node: ts.BinaryExpression, lhsStr: string, rhsStr: string, wrap = false) {
@@ -129,7 +131,18 @@ export function compileBinaryExpression(state: CompilerState, node: ts.BinaryExp
 	let lhsStr: string;
 	let rhsStr: string;
 
-	if (!isEqualsOperation) {
+	if (isEqualsOperation) {
+		const isLhsMethod = isDefinedAsMethod(lhs);
+		if (isLhsMethod !== undefined && isLhsMethod !== isMethodDeclaration(rhs)) {
+			throw new CompilerError(
+				`Attempted to set a ${isLhsMethod ? "method" : "callback"} variable to a ${
+					isLhsMethod ? "callback" : "method"
+				}.`,
+				node,
+				CompilerErrorType.MixedMethodSet,
+			);
+		}
+	} else {
 		checkNonAny(lhs);
 		checkNonAny(rhs);
 	}